# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pilot Web UI API",
    "description": "BFF server API for Pilot Web UI — chat sessions, platform channels, model management, skills, memory, logs, and terminal.",
    "version": "0.2.6"
  },
  "servers": [
    { "url": "http://localhost:8648", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/pilot/logs": {
      "get": {
        "tags": ["Logs"],
        "summary": "List log files",
        "operationId": "listLogs",
        "responses": {
          "200": { "description": "Log file names", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LogListResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/logs/{name}": {
      "get": {
        "tags": ["Logs"],
        "summary": "Read log file",
        "operationId": "readLog",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Log file name" },
          { "name": "lines", "in": "query", "schema": { "type": "integer", "default": 100 }, "description": "Number of lines" },
          { "name": "level", "in": "query", "schema": { "type": "string", "enum": ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"] }, "description": "Filter by level" },
          { "name": "session", "in": "query", "schema": { "type": "string" }, "description": "Filter by session ID" },
          { "name": "since", "in": "query", "schema": { "type": "string" }, "description": "Filter entries since timestamp" }
        ],
        "responses": {
          "200": { "description": "Log entries", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LogEntriesResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to read", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/profiles": {
      "get": {
        "tags": ["Profiles"],
        "summary": "List all profiles",
        "operationId": "listProfiles",
        "responses": {
          "200": { "description": "Profile list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileListResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      },
      "post": {
        "tags": ["Profiles"],
        "summary": "Create a new profile",
        "operationId": "createProfile",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProfileRequest" } } } },
        "responses": {
          "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } } } } } },
          "400": { "description": "Missing profile name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to create", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/profiles/active": {
      "put": {
        "tags": ["Profiles"],
        "summary": "Switch active profile",
        "description": "Stops gateway, switches profile, then restarts gateway. Supports WSL and Windows.",
        "operationId": "switchProfile",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SwitchProfileRequest" } } } },
        "responses": {
          "200": { "description": "Switched", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } } } } } },
          "400": { "description": "Missing profile name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to switch", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/profiles/import": {
      "post": {
        "tags": ["Profiles"],
        "summary": "Import profile from archive",
        "operationId": "importProfile",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportProfileRequest" } } } },
        "responses": {
          "200": { "description": "Imported", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } } } } } },
          "400": { "description": "Missing archive path", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to import", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/profiles/{name}": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Get profile details",
        "operationId": "getProfile",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Profile details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileDetailResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Profile not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Failed to get profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      },
      "delete": {
        "tags": ["Profiles"],
        "summary": "Delete a profile",
        "description": "Cannot delete the 'default' profile.",
        "operationId": "deleteProfile",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Cannot delete the default profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to delete", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/profiles/{name}/rename": {
      "post": {
        "tags": ["Profiles"],
        "summary": "Rename a profile",
        "operationId": "renameProfile",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RenameProfileRequest" } } } },
        "responses": {
          "200": { "description": "Renamed", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Missing new_name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to rename", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/profiles/{name}/export": {
      "post": {
        "tags": ["Profiles"],
        "summary": "Export profile to archive",
        "operationId": "exportProfile",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExportProfileRequest" } } } },
        "responses": {
          "200": { "description": "Exported", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to export", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
...
    "/api/pilot/weixin/qrcode": {
      "get": {
        "tags": ["Weixin"],
        "summary": "Fetch WeChat QR code",
        "operationId": "getWeixinQrcode",
        "responses": {
          "200": { "description": "QR code data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WeixinQrcodeResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to fetch QR code", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/weixin/qrcode/status": {
      "get": {
        "tags": ["Weixin"],
        "summary": "Poll QR scan status",
        "operationId": "getWeixinQrcodeStatus",
        "parameters": [{ "name": "qrcode", "in": "query", "required": true, "schema": { "type": "string" }, "description": "QR code identifier" }],
        "responses": {
          "200": { "description": "Scan status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WeixinQrcodeStatusResponse" } } } },
          "400": { "description": "Missing qrcode parameter", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to poll", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/weixin/save": {
      "post": {
        "tags": ["Weixin"],
        "summary": "Save WeChat credentials",
        "operationId": "saveWeixinCredentials",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveWeixinRequest" } } } },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Missing required fields", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to save", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Auth token from dist/server/data/.token or AUTH_TOKEN env var"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
...
      "LogListResponse": {
        "type": "object",
        "properties": {
          "files": { "type": "array", "items": { "type": "string" }, "description": "Log file names" }
        }
      },
      "LogEntriesResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "timestamp": { "type": "string" },
                    "level": { "type": "string", "enum": ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"] },
                    "logger": { "type": "string" },
                    "message": { "type": "string" },
                    "raw": { "type": "string" }
                  }
                },
                { "type": "null", "description": "Unparseable line (continuation, traceback, etc.)" }
              ]
            }
          }
        }
      },
      "WeixinQrcodeResponse": {
        "type": "object",
        "properties": {
          "qrcode": { "type": "string", "description": "QR code identifier" },
          "qrcode_url": { "type": "string", "description": "QR code image (base64 data URL)" }
        }
      },
      "WeixinQrcodeStatusResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["wait", "confirmed"] },
          "account_id": { "type": "string", "description": "Bot account ID (when confirmed)" },
          "token": { "type": "string", "description": "Bot token (when confirmed)" },
          "base_url": { "type": "string", "description": "Base URL (when confirmed)" }
        }
      },
      "SaveWeixinRequest": {
        "type": "object",
        "properties": {
          "account_id": { "type": "string" },
          "token": { "type": "string" },
          "base_url": { "type": "string" }
        },
        "required": ["account_id", "token"]
      },
...
      "ProfileListResponse": {
        "type": "object",
        "properties": {
          "profiles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "active": { "type": "boolean" },
                "model": { "type": "string" },
                "gateway": { "type": "string" },
                "alias": { "type": "string" }
              }
            }
          }
        }
      },
      "ProfileDetailResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "path": { "type": "string" },
              "model": { "type": "string" },
              "provider": { "type": "string" },
              "gateway": { "type": "string" },
              "skills": { "type": "integer" },
              "hasEnv": { "type": "boolean" },
              "hasSoulMd": { "type": "boolean" }
            }
          }
        }
      },
      "CreateProfileRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Profile name (lowercase alphanumeric only)" },
          "clone": { "type": "boolean", "description": "Clone from current profile's config, .env, and soul.md" }
        },
        "required": ["name"]
      },
      "SwitchProfileRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Profile name to switch to" }
        },
        "required": ["name"]
      },
      "RenameProfileRequest": {
        "type": "object",
        "properties": {
          "new_name": { "type": "string", "description": "New profile name" }
        },
        "required": ["new_name"]
      },
      "ExportProfileRequest": {
        "type": "object",
        "properties": {
          "output": { "type": "string", "description": "Output archive path (optional)" }
        }
      },
      "ImportProfileRequest": {
        "type": "object",
        "properties": {
          "archive": { "type": "string", "description": "Path to archive file" },
          "name": { "type": "string", "description": "Profile name after import (optional)" }
        },
        "required": ["archive"]
      }
    }
  }
}